requirements.txt
google-generativeai==0.7.2
SpeechRecognition==3.10.4
PyAudio==0.2.14
python-dotenv==1.0.1

